requirements.txt
Flask==2.0.2
numpy==1.21.1
opencv-python==4.5.3.56
face-recognition==1.3.0
Pillow==8.3.2